% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{create_dt_output}
\alias{create_dt_output}
\title{Create DataTable output}
\usage{
create_dt_output(ns, values, value_name, link_column = "rownames")
}
\arguments{
\item{ns}{Namespace function}

\item{values}{Reactive values}

\item{value_name}{Name of the value to display}

\item{link_column}{Column to create links for}
}
\value{
A DataTable output
}
\description{
Create a DataTable output for displaying data
}
\keyword{internal}
