% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{create_gene_list_handler}
\alias{create_gene_list_handler}
\title{Create gene list handler}
\usage{
create_gene_list_handler(list_num, input, values, ns, read1stCol)
}
\arguments{
\item{list_num}{Number of the gene list (1-4)}

\item{input}{Shiny input object}

\item{values}{Reactive values}

\item{ns}{Namespace function}

\item{read1stCol}{Function to read the first column of a file}
}
\value{
A reactive function for the gene list
}
\description{
Create a reactive and observer for handling gene list uploads
}
\keyword{internal}
