% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{create_goterm_heatmap}
\alias{create_goterm_heatmap}
\title{Create GO term heatmap}
\usage{
create_goterm_heatmap(topgo_data, selected_row, values, annoSpecies_df)
}
\arguments{
\item{topgo_data}{TopGO data containing GO terms and genes}

\item{selected_row}{Selected row index from the data}

\item{values}{Reactive values containing dds_obj, cur_species, and cur_type}

\item{annoSpecies_df}{Data frame with species information}
}
\value{
A heatmaply plot or NULL if no row is selected
}
\description{
Create a heatmap for selected GO terms and genes
}
\keyword{internal}
