% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{create_ui_output}
\alias{create_ui_output}
\title{Create UI output}
\usage{
create_ui_output(ns, values, value_name, title, output_name)
}
\arguments{
\item{ns}{Namespace function}

\item{values}{Reactive values}

\item{value_name}{Name of the value to display}

\item{title}{Title for the output}

\item{output_name}{Name of the output}
}
\value{
A UI output
}
\description{
Create a UI output for displaying data tables
}
\keyword{internal}
