% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_gene_lists}
\alias{generate_gene_lists}
\title{Generate DE gene lists}
\usage{
generate_gene_lists(res_obj, FDR = 0.05, annotation_obj = NULL)
}
\arguments{
\item{res_obj}{DESeqResults object}

\item{FDR}{False discovery rate threshold}

\item{annotation_obj}{Annotation data frame (optional)}
}
\value{
A list with UP, DOWN, and UPDOWN gene lists
}
\description{
Extract upregulated, downregulated, and all DE genes from results.
}
