% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Get gene list}
\usage{
get_gene_list(list_name, values)
}
\arguments{
\item{list_name}{Name of the gene list to retrieve}

\item{values}{Reactive values containing the gene lists}
}
\value{
The requested gene list or NULL if not found
}
\description{
Retrieve a gene list from the provided values based on the list name
}
\keyword{internal}
