% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{mat_rowscale}
\alias{mat_rowscale}
\title{Row scaling function}
\usage{
mat_rowscale(x)
}
\arguments{
\item{x}{A matrix or data frame to be scaled}
}
\value{
A scaled matrix with the same dimensions as the input
}
\description{
Scale matrix rows by subtracting the mean and dividing by the standard deviation
}
\keyword{internal}
