% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEIN.R
\name{multiAxPCA}
\alias{multiAxPCA}
\title{Multi-dimensional PCA plot}
\usage{
multiAxPCA(
  object,
  intgroup = "condition",
  ntop = 500,
  returnData = FALSE,
  pc1 = 1,
  pc2 = 2
)
}
\arguments{
\item{object}{A DESeqTransform object}

\item{intgroup}{Character vector of interesting groups}

\item{ntop}{Number of top genes to use}

\item{returnData}{Logical, whether to return data instead of plot}

\item{pc1}{First principal component to plot}

\item{pc2}{Second principal component to plot}
}
\value{
Either a ggplot object or PCA data
}
\description{
Creates a PCA plot with customizable dimensions
}
