% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read1stCol}
\alias{read1stCol}
\title{Read the first column from a file}
\usage{
read1stCol(fileName, dds_obj)
}
\arguments{
\item{fileName}{Path to the file to read}

\item{dds_obj}{A DESeqDataSet object used for filtering rows}
}
\value{
A data frame with the first column, filtered based on the dds_obj
}
\description{
Reads the first column from a file, guessing the separator automatically and
filtering rows based on a DESeqDataSet object.
}
