% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_countmatrix}
\alias{read_countmatrix}
\title{Read count matrix from file}
\usage{
read_countmatrix(filepath, sepguesser_function = sepguesser)
}
\arguments{
\item{filepath}{Path to the file to read}

\item{sepguesser_function}{Function to use for guessing the separator (defaults to sepguesser)}
}
\value{
A matrix with count data
}
\description{
Reads a count matrix from a file, guessing the separator automatically.
}
