% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesignatures.R
\name{read_gmt}
\alias{read_gmt}
\title{Read in a GMT file}
\usage{
read_gmt(gmtfile)
}
\arguments{
\item{gmtfile}{A character value, containing the location of the GMT formatted
file. It can also be a file found online}
}
\value{
A list of vectors, one for each pathway in the GMT file.
}
\description{
Returns a list of pathways from a GMT file.
}
\examples{
# this example shows how the function works with a simple example
# create a temporary GMT file for demonstration
tmp_gmt <- tempfile(fileext = ".gmt")
cat("PATHWAY1\tsource1\tGENE1\tGENE2\tGENE3\n", file = tmp_gmt)
cat("PATHWAY2\tsource2\tGENE4\tGENE5\n", file = tmp_gmt, append = TRUE)
mysigs <- read_gmt(tmp_gmt)
head(mysigs)
unlink(tmp_gmt)  # clean up
# see how the gene identifiers are encoded as ENTREZ id
}
