% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata from file}
\usage{
read_metadata(filepath, sepguesser_function = sepguesser)
}
\arguments{
\item{filepath}{Path to the file to read}

\item{sepguesser_function}{Function to use for guessing the separator (defaults to sepguesser)}
}
\value{
A data frame with metadata
}
\description{
Reads metadata from a file, guessing the separator automatically.
}
