% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\docType{package}
\name{utilities_internal}
\alias{utilities_internal}
\title{Internal Utility Functions}
\description{
These functions are used internally by the WEIN package and are not meant
to be called directly by users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baj12/WEIN}
  \item \url{https://baj12.github.io/WEIN/}
  \item Report bugs at \url{https://github.com/baj12/WEIN/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Bernd Jagla \email{bernd.jagla@pasteur.fr} (Developer of immunology-focused version) [contributor]
}

}
\keyword{internal}
