% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEIN.R
\name{WEIN-package}
\alias{WEIN-package}
\title{WEIN: Web-based Engine for Interactive Next-generation sequencing analysis}
\arguments{
\item{dds_obj}{A \code{\link{DESeqDataSet}} object. If not provided, then a
\code{countmatrix} and a \code{expdesign} need to be provided. If none of
the above is provided, it is possible to upload the data during the
execution of the Shiny App}

\item{res_obj}{A \code{\link{DESeqResults}} object. If not provided, it can
be computed during the execution of the application}

\item{annotation_obj}{A \code{data.frame} object, with row.names as gene
identifiers (e.g. ENSEMBL ids) and a column, \code{gene_name}, containing
e.g. HGNC-based gene symbols. If not provided, it can be constructed during
the execution via the org.eg.XX.db packages - these need to be installed}

\item{countmatrix}{A count matrix, with genes as rows and samples as columns.
If not provided, it is possible to upload the data during the execution of
the Shiny App}

\item{expdesign}{A \code{data.frame} containing the info on the covariates
of each sample. If not provided, it is possible to upload the data during the
execution of the Shiny App}

\item{gene_signatures}{A list of vectors, one for each pathway/signature. This
is for example the output of the \code{\link{read_gmt}} function. The provided
object can also be replaced during runtime in the dedicated upload widget.}

\item{dds_design}{A character vector specifying the design formula for DESeq2.
This will be used when creating the DESeqDataSet object.}

\item{choose_expfac}{A character vector specifying the experimental factor to build the contrast.
This will be used when extracting results from the DESeqDataSet object.}

\item{state_file}{An optional path to a .RData file containing saved application state.
If provided, the application state will be restored from this file.}
}
\value{
A Shiny App is launched for interactive data exploration and
differential expression analysis
}
\description{
WEIN makes bulk RNAseq data analysis interactive, easy and reproducible.
This function launches the main application included in the package.
}
\section{Debugging Options}{

The application includes debugging output that can be enabled/disabled
using the global option \code{wein.debug}. By default, debugging output
is disabled to reduce console clutter.

To enable debugging output, set the option before launching the app:
\code{options(wein.debug = TRUE)}

When enabled, detailed timing information and debug messages will be
printed to the console during application execution.
}

\examples{
# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n=100, m=8)
cm <- counts(dds)
cd <- colData(dds)

# with the well known airway package...
library(airway)
data(airway)
airway
dds_airway <- DESeq2::DESeqDataSetFromMatrix(assay(airway),
                                             colData = colData(airway),
                                             design=~cell+dex)
\dontrun{

WEIN()
WEIN(dds)
WEIN(dds_airway)

dds_airway <- DESeq2::DESeq(dds_airway)
res_airway <- DESeq2::results(dds_airway)
WEIN(dds_airway, res_airway)
}

# Restore from a saved state file
\dontrun{
WEIN(state_file = "my_analysis_state.RData")
}

# Save current state from within the app:
# 1. Click on "Save State as .RData" button in the header
# 2. This creates a timestamped .RData file with your current analysis state
# 3. Later, restore the session with:
#    WEIN(state_file = "WEINState_2023-12-01_14-30-45.RData")
#    or with a pre-selected experimental factor:
#    WEIN(state_file = "WEINState_2023-12-01_14-30-45.RData", choose_expfac = "condition")


}
