% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEIN-pkg.R
\docType{package}
\name{WEIN-pkg}
\alias{WEIN-pkg}
\title{WEIN: Web-based Engine for Interactive Next-generation sequencing analysis}
\description{
WEIN makes bulk RNAseq data analysis interactive, easy and reproducible.
The analysis of RNA-seq datasets is guided by the Shiny app as main component of
the package, which also provides a wide set of functions to efficiently extract
information from the existing data. The app can be also deployed on a Shiny
server, to allow its usage without any installation on the user's side.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baj12/WEIN}
  \item \url{https://baj12.github.io/WEIN/}
  \item Report bugs at \url{https://github.com/baj12/WEIN/issues}
}

}
\author{
Bernd Jagla \email{bernd.jagla@pasteur.fr}, 2025

Maintainer: Bernd Jagla \email{bernd.jagla@pasteur.fr}
}
\keyword{internal}
