% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_components.R
\docType{package}
\name{WEIN-server}
\alias{WEIN-server}
\title{Server Components for WEIN}
\description{
Contains all server-related functions for the WEIN Shiny application.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baj12/WEIN}
  \item \url{https://baj12.github.io/WEIN/}
  \item Report bugs at \url{https://github.com/baj12/WEIN/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Bernd Jagla \email{bernd.jagla@pasteur.fr} (Developer of immunology-focused version) [contributor]
}

}
\keyword{internal}
