% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEIN.R
\name{WEIN}
\alias{WEIN}
\title{Main function to launch the WEIN Shiny application}
\usage{
WEIN(
  dds_obj = NULL,
  res_obj = NULL,
  annotation_obj = NULL,
  countmatrix = NULL,
  expdesign = NULL,
  gene_signatures = NULL,
  dds_design = NULL,
  choose_expfac = NULL,
  choose_expfac2 = NULL,
  cur_species = NULL,
  cur_type = NULL,
  state_file = NULL
)
}
\arguments{
\item{dds_obj}{A DESeqDataSet object}

\item{res_obj}{A DESeqResults object}

\item{annotation_obj}{A data frame with gene annotations}

\item{countmatrix}{A count matrix}

\item{expdesign}{A data frame with experimental design}

\item{gene_signatures}{A list of gene signatures}

\item{dds_design}{A character vector specifying the design formula for DESeq2}

\item{choose_expfac}{A character vector specifying the experimental factor to build the contrast (numerator)}

\item{choose_expfac2}{A character vector specifying the experimental factor to build the contrast (denominator)}

\item{state_file}{An optional path to a .RData file containing saved application state}
}
\value{
A Shiny application
}
\description{
Launches the interactive differential expression analysis application
}
