% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_components.R
\name{WEIN_server}
\alias{WEIN_server}
\title{Create the server for WEIN}
\usage{
WEIN_server(
  app_data = list(dds_obj = NULL, res_obj = NULL, annotation_obj = NULL, countmatrix =
    NULL, expdesign = NULL, gene_signatures = NULL, dds_design = NULL, choose_expfac =
    NULL, choose_expfac2 = NULL, cur_species = NULL, cur_type = NULL)
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{dds_obj}{DESeqDataSet object}

\item{res_obj}{DESeqResults object}

\item{annotation_obj}{Annotation data frame}

\item{countmatrix}{Count matrix}

\item{expdesign}{Experimental design data frame}

\item{gene_signatures}{Gene signatures list}

\item{dds_design}{Design formula for DESeq2}

\item{choose_expfac}{Experimental factor to build the contrast (numerator)}

\item{choose_expfac2}{Experimental factor to build the contrast (denominator)}

\item{cur_species}{Current species for annotation}

\item{cur_type}{Current ID type for annotation}
}
\value{
A Shiny server definition
}
\description{
Defines the complete server logic for the WEIN Shiny application.
}
